Renamed integer.rs into uint128.rs
commit-id:82da9c43
