+++ crates/sierra/src/extensions/core.rs
@@ -11,9 +11,9 @@ use super::modules::boxing::{BoxLibFunc, BoxType};
 use super::modules::felt::{FeltLibFunc, FeltType};
 use super::modules::function_call::FunctionCallLibFunc;
 use super::modules::gas::{GasBuiltinType, GasLibFunc};
-use super::modules::integer::{Uint128LibFunc, Uint128Type};
 use super::modules::mem::MemLibFunc;
 use super::modules::non_zero::{NonZeroType, UnwrapNonZeroLibFunc};
+use super::modules::uint128::{Uint128LibFunc, Uint128Type};
 use super::modules::unconditional_jump::UnconditionalJumpLibFunc;
 use super::pedersen::{PedersenLibFunc, PedersenType};
 use super::range_check::RangeCheckType;
